Use users controller
